gdk/frameclockidle: avoid narrowing timing precision
Keep smooth phase modulo arithmetic in gint64 so frame-clock
timestamps are not narrowed through int. It would take roughly 35
minutes for this to happen, but it is an easy cleanup.

+++ gdk/gdkframeclockidle.c
@@ -454,8 +454,9 @@ gdk_frame_clock_flush_idle (void *data)
  *
  * The built-in `%` operator returns the former, positive_modulo() returns the latter.
  */
-static int
-positive_modulo (int i, int n)
+static gint64
+positive_modulo (gint64 i,
+                 gint64 n)
 {
   return (i % n + n) % n;
 }
